clean up shared router/tagger ui stuff
